+++ CentroDeportivo.Aplicacion/Entidades/Devolucion.cs
@@ -9,11 +9,11 @@ namespace CentroDeportivo.Aplicacion.Entidades
     public class Devolucion
     {
         public int Id { get; private set; }
-        public double MontoADevolver { get; set; }
+        public decimal MontoADevolver { get; set; }
         public DevolucionEstado Estado { get; set; }
         public int Id_Usuario { get; set; }
         public Usuario Usuario { get; set; }
-        public int Id_Reserva {  get; set; }
+        public int Id_Reserva { get; set; }
         public Reserva Reserva { get; set; }
         public DateTime FechaGeneracion { get; set; }
         public DateTime? FechaDevolucion { get; set; }
